Fix printf format specifiers for windows compatibility using macros from <inttypes.h>
For explanation, see e.g.
https://en.cppreference.com/w/cpp/header/cinttypes
https://stackoverflow.com/questions/16859500/mmh-who-are-you-priu64

+++ src/os/windows/syncio_file.c
@@ -357,9 +357,9 @@ int64_t GWEN_SyncIo_File_Seek(GWEN_SYNCIO *sio, int64_t pos, GWEN_SYNCIO_FILE_WH
     rv=GetLastError();
     if (rv!=NO_ERROR) {
       DBG_ERROR(GWEN_LOGDOMAIN,
-                "Error seeking in file[%s] at %llu: %d",
+                "Error seeking in file[%s] at %" PRIi64 ": %d",
                 xio->path,
-                (unsigned long long) pos,
+                (int64_t) pos,
                 (int)rv);
       return GWEN_ERROR_IO;
     }
